Changed the way training models works -> 1 model per person

+++ customer_pipeline/src/train_classifier.py
@@ -1,17 +1,24 @@
 #!/usr/bin/env python
 import tensorflow as tf
-import numpy as np
 import rospy
 import facenet
 import argparse
-import math
 import pickle
 import time
-import datetime
 import os
+import numpy as np
 from sklearn.svm import SVC
 from std_msgs.msg import String
-from datetime import datetime
+
+
+def read_paths_from_folder(folder, paths):
+    nrof_paths = 0
+    for img in os.listdir(folder):
+        if '.png' in img:
+            nrof_paths += 1
+            paths.append(os.path.join(folder, img))
+
+    return nrof_paths
 
 
 class TrainClassifier:
@@ -19,9 +26,15 @@ class TrainClassifier:
 
     def __init__(self, args):
         self.folder = args.data_dir
+        self.output = args.output_dir
         self.model = args.model
         self.topic = args.topic
         self.gpu_options = tf.GPUOptions(per_process_gpu_memory_fraction=args.gpu_memory_fraction)
+
+        try:
+            self.person_embs = np.load(os.path.join(self.output, 'embeddings.npy'))
+        except IOError:
+            self.person_embs = {}
 
         with tf.Graph().as_default():
             with tf.Session(
@@ -38,40 +51,45 @@ class TrainClassifier:
 
                 self.sub = rospy.Subscriber(self.topic, String, self.callback)
 
+                rospy.spin()
+
     def callback(self, data):
-        print('Training classifier')
+        print('Training classifier(s)')
+        for person in os.listdir(self.folder):
+            if person != 'Unknown':
+                if not os.path.exists(os.path.join(self.output, '{}.pkl'.format(person))) \
+                        or time.time() - os.path.getmtime(os.path.join(self.output, '{}.pkl'.format(person))) \
+                                > (30 * 24 * 60 * 60):
+                    print('Creating classifier for: {}'.format(person))
+                    paths = []
+                    nrof_paths = read_paths_from_folder(os.path.join(self.folder, person), paths)
+                    labels = [0] * nrof_paths
+                    nrof_paths = read_paths_from_folder(os.path.join(self.folder, 'Unknown'), paths)
+                    labels += [1] * nrof_paths
 
-        model = str(datetime.strftime(datetime.now(), '%Y%m%d-%H%M%S')) + '.pkl'
+                    classifier_filename_exp = os.path.expanduser(
+                        os.path.join(self.output, '{}.pkl'.format(person)))
 
-        classifier_filename_exp = os.path.expanduser('/home/maarten/Documents/ProjectIris/model/' + model)
+                    images = facenet.load_data(paths, False, False, 160)
+                    feed_dict = {self.images_placeholder: images, self.phase_train_placeholder: False}
+                    emb_array = self.sess.run(self.embeddings, feed_dict=feed_dict)
 
-        dataset = facenet.get_dataset(self.folder)
-        paths, labels = facenet.get_image_paths_and_labels(dataset)
+                    model = SVC(kernel='linear', probability=True)
+                    model.fit(emb_array, labels)
 
-        print('Number of classes: %d' % len(dataset))
-        print('Number of images: %d' % len(paths))
+                    name = person.replace('_', ' ')
+                    class_names = [name, 'Unknown']
 
-        nrof_images = len(paths)
-        nrof_batches_per_epoch = int(math.ceil(1.0 * nrof_images / self.BATCH_SIZE))
-        emb_array = np.zeros((nrof_images, self.embedding_size))
-        for i in range(nrof_batches_per_epoch):
-            start_index = i * self.BATCH_SIZE
-            end_index = min((i + 1) * self.BATCH_SIZE, nrof_images)
-            paths_batch = paths[start_index:end_index]
-            images = facenet.load_data(paths_batch, False, False, 160)
-            feed_dict = {self.images_placeholder: images, self.phase_train_placeholder: False}
-            emb_array[start_index:end_index, :] = self.sess.run(self.embeddings, feed_dict=feed_dict)
+                    with open(classifier_filename_exp, 'wb') as outfile:
+                        pickle.dump((model, class_names), outfile)
 
-        model = SVC(kernel='linear', probability=True)
-        model.fit(emb_array, labels)
+                    if name in self.person_embs:
+                        self.person_embs[name] = emb_array[0]
+                    else:
+                        self.person_embs.update({name: emb_array[0]})
 
-        class_names = [cls.name.replace('_', ' ') for cls in dataset]
-
-        with open(classifier_filename_exp, 'wb') as outfile:
-            pickle.dump((model, class_names), outfile)
-
-        print('Saved classifier model to file "%s"' % classifier_filename_exp)
-        self.pub.publish(classifier_filename_exp)
+                    print('Saved classifier model to file "%s"' % classifier_filename_exp)
+                    self.pub.publish(classifier_filename_exp)
 
 
 if __name__ == '__main__':
@@ -82,6 +100,8 @@ if __name__ == '__main__':
                         help='Path to a model protobuf (.pb) file.')
     parser.add_argument('data_dir', type=str,
                         help='Path to dir containing the aligned person photos.')
+    parser.add_argument('output_dir', type=str,
+                        help='Path to where the classifier (.pkl) files will be saved.')
     parser.add_argument('--topic', type=str,
                         help='The ros topic to listen to.', default='/train_command')
     parser.add_argument('--gpu_memory_fraction', type=float,
@@ -90,6 +110,3 @@ if __name__ == '__main__':
     args = parser.parse_args()
 
     trainer = TrainClassifier(args)
-
-    while True:
-        time.sleep(5)
